requirements.txt
python-dotenv
twitchio
tortoise-orm
cacheout
regex
humanize
